+++ Package.swift
@@ -17,14 +17,6 @@ let package = Package(
         .library(
             name: "PocketStorage",
             targets: ["PocketStorage"]
-        ),
-        .library(
-            name: "PocketStorageObservation",
-            targets: ["PocketStorageObservation"]
-        ),
-        .library(
-            name: "PocketStorageUI",
-            targets: ["PocketStorageUI"]
         )
     ],
     targets: [
@@ -39,27 +31,13 @@ let package = Package(
             name: "PocketStorage",
             path: "Sources/PocketStorage"
         ),
-        .target(
-            name: "PocketStorageObservation",
-            dependencies: ["PocketStorage"],
-            path: "Sources/PocketStorageObservation"
-        ),
-        .target(
-            name: "PocketStorageUI",
-            dependencies: ["PocketStorage"],
-            path: "Sources/PocketStorageUI"
-        ),
         .testTarget(
             name: "PocketUITests",
             dependencies: ["PocketUI"]
         ),
         .testTarget(
             name: "PocketStorageTests",
-            dependencies: ["PocketStorage", "PocketStorageObservation"]
-        ),
-        .testTarget(
-            name: "PocketStorageUITests",
-            dependencies: ["PocketStorage", "PocketStorageUI"]
+            dependencies: ["PocketStorage"]
         ),
     ]
 )