+++ Tools/DocC/Sources/DocCTools/DocCTools.swift
@@ -1,3 +1,5 @@
 import PocketUI
+import PocketStorage
 
 private let _docCToolsUsesPocketUI = GalleryLayout.ContentDisplayMode.self
+private let _docCToolsUsesPocketStorage = PocketStore.self